% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmatplot-package.R, R/ggmatplot.R
\docType{ggmatplot}
\name{ggmatplot}
\alias{ggmatplot}
\title{ggmatplot}
\usage{
ggmatplot(x, y, color = NULL, shape = NULL, linetype = NULL,
  xlim = c(NA, NA), ylim = c(NA, NA), log = "", main = NULL,
  xlab = NULL, ylab = NULL, legend_label = NULL,
  legend_title = NULL, geom = "point", asp = NA)
}
\arguments{
\item{x, y}{vectors or matrices of data for plotting. The number of rows should match. If one of them are missing, the other is taken as y and an x vector of 1:n is used. Missing values (NAs) are allowed.}

\item{color}{vector of colors and they are used cyclically.}

\item{shape}{vector of shapes and they are used cyclically.}

\item{linetype}{vector of linetypes colors  and they are used cyclically.}

\item{xlim, ylim}{ranges of x and y axes, as in [plot()].}

\item{log}{Which variables to log transform ("x", "y", or "xy")}

\item{main, xlab, ylab}{Character vector giving plot title,
x axis label, and y axis label respectively.}

\item{legend_label}{Character vector giving legend tables for different groups.}

\item{legend_title}{Character giving legend title}

\item{geom}{Character vector specifying geom(s) to draw. Defaults to
"point". Other options are "line" or c("point","line").}

\item{asp}{The y/x aspect ratio}
}
\description{
This package was created as a way to make a matplot with ggplot version

`ggmatplot` is a quick way to make a simplier and similar plot as what `matplot` does but with a ggplot version.
}
\examples{
# Define a data set
iris_sub <- subset(iris, Species == "setosa")
# Use default legend and axis
ggmatplot(iris_sub[,c(1,3)], iris_sub[,c(2,4)])
# Modify legend label and axis
ggmatplot(iris_sub[,c(1,3)], iris_sub[,c(2,4)], shape = c("s","S"), legend_label =  c("Sepal","Petal"), legend_title = "", xlab = "Length", ylab="Width")
}
