% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmatplot.R
\name{ggmatplot}
\alias{ggmatplot}
\title{ggmatplot}
\usage{
ggmatplot(
  x,
  y,
  color = NULL,
  shape = NULL,
  linetype = NULL,
  fill = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  log = "",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_label = NULL,
  legend_title = NULL,
  plot_type = "point",
  asp = NA,
  ...
)
}
\arguments{
\item{x, y}{Vectors or matrices of data for plotting.
\itemize{
\item The number of rows of \code{x} and \code{y} should be the same.
\item Missing values (NAs) are allowed.
\item If either \code{x} or \code{y} is missing, the other is used as \code{y} and a vector of
\code{1:n} is used as \code{x}.
}}

\item{color, fill}{Vectors of colors. Defining only one of them will update
both \code{color} and \code{fill} aesthetics of the plot by default, unless they are
both defined simultaneously.
\itemize{
\item The number of colors should match the higher number of columns of
matrices \code{x} or \code{y}, and will correspond to each of those columns.
\item If only a single color is given, the same color will be used for all columns.
}}

\item{shape, linetype}{A vector of shapes or line types respectively.
\itemize{
\item The number of shapes/line types should match the higher number of columns of
matrices \code{x} or \code{y}, and will correspond to each of those columns.
\item If only a single shape/line type is given, the same shape/line type will be
used for all columns.
}}

\item{xlim, ylim}{Ranges of x and y axes. Each of them should be a two element
vector specifying the lower and upper limits of the scale. If the larger
value is given first, the scale will be reversed. If one of the limits is
given as \code{NA}, the corresponding limit from the range of data will be used.}

\item{log}{A string defining which axes to transform into a log scale.
(\code{x}, \code{y} or \code{xy})}

\item{main, xlab, ylab, legend_title}{Strings to update plot title, x axis label,
y axis label and legend title respectively.}

\item{legend_label}{A list of strings, to rename the legend labels.}

\item{plot_type}{A string specifying the type of plot to draw. Possible plot
types are \code{point}, \code{line}, \code{both}(point + line), \code{density}, \code{histogram},
\code{boxplot}, \code{dotplot}, \code{errorplot}, \code{violin}, and \code{ecdf}. Default plot_type is
\code{point}.}

\item{asp}{The y/x aspect ratio.}

\item{...}{Other arguments passed on to the plot. They can be used to set an
aesthetic to a fixed value, like \code{alpha = 0.4} or \code{size = 2}. They may also
be parameters passed on to the corresponding \href{https://ggplot2.tidyverse.org/reference/}{geoms}.}
}
\description{
\code{ggmatplot} is a quick and easy way of plotting two matrices against each
other. Can also be used as an alternative to pivoting data into long format
before plotting, by using the index column(s) and the columns to be plotted
against the index as the two matrix inputs for \code{ggmatplot()}.
}
\examples{

# Define a data set
iris_sub <- subset(iris, Species == "setosa")
ggmatplot(iris_sub[, c(1, 3)], iris_sub[, c(2, 4)])
# Modify legend label and axis
ggmatplot(iris_sub[, c(1, 3)], iris_sub[, c(2, 4)], shape = c(4, 6), legend_label = c("Sepal", "Petal"), legend_title = "", xlab = "Length", ylab = "Width")


}
